Refactors application layouts
