--- src/lib/components/layouts/MainLayout.svelte
@@ -1,283 +1,0 @@
-<script lang="ts">
-  import type { Writable } from "svelte/store";
-  import type { Org, User, Flatpage } from "$lib/api/types";
-
-  import { _ } from "svelte-i18n";
-  import { page } from "$app/stores";
-
-  import { getContext } from "svelte";
-  import { SidebarCollapse16, SidebarExpand16 } from "svelte-octicons";
-
-  import Button from "../common/Button.svelte";
-  import Flex from "../common/Flex.svelte";
-  import SignedIn from "../common/SignedIn.svelte";
-  import UserMenu from "../accounts/UserMenu.svelte";
-  import OrgMenu from "../navigation/OrgMenu.svelte";
-
-  import { SIGN_IN_URL } from "@/config/config";
-  import Breadcrumbs from "../navigation/Breadcrumbs.svelte";
-  import LanguageMenu from "../navigation/LanguageMenu.svelte";
-  import HelpMenu from "../navigation/HelpMenu.svelte";
-  import Toaster from "./Toaster.svelte";
-  import UserFeedback from "../forms/UserFeedback.svelte";
-  import Portal from "./Portal.svelte";
-  import Modal from "./Modal.svelte";
-  import TipOfDay from "../common/TipOfDay.svelte";
-
-  const me = getContext<Writable<User>>("me");
-  const org = getContext<Writable<Org>>("org");
-  const tipOfDay = getContext<Flatpage>("tipOfDay");
-
-  const user_orgs = getContext<Writable<Promise<Org[]>>>("user_orgs");
-  const org_users = getContext<Writable<Promise<User[]>>>("org_users");
-
-  let panel: "navigation" | "action" | null = null;
-
-  function closePanel() {
-    panel = null;
-  }
-
-  function openPanel(name: typeof panel): () => void {
-    return () => {
-      panel = name;
-    };
-  }
-
-  let feedbackOpen = false;
-</script>
-
-<div class="container">
-  {#if tipOfDay}<TipOfDay message={tipOfDay.content} />{/if}
-  <header>
-    {#if $$slots.navigation}
-      <div class="small openPane">
-        <Button
-          minW={false}
-          ghost
-          mode="primary"
-          on:click={openPanel("navigation")}
-        >
-          <SidebarCollapse16 />
-        </Button>
-      </div>
-    {/if}
-    <slot name="breadcrumbs">
-      <Breadcrumbs trail={$page.data.breadcrumbs} />
-    </slot>
-    <SignedIn>
-      <Flex>
-        {#await Promise.all([$user_orgs, $org_users]) then [orgs, users]}
-          <OrgMenu active_org={$org} {orgs} {users} />
-        {/await}
-        <UserMenu user={$me} />
-      </Flex>
-      <Button slot="signedOut" mode="primary" href={SIGN_IN_URL}>
-        {$_("authSection.user.signIn")}
-      </Button>
-    </SignedIn>
-    <LanguageMenu />
-    <HelpMenu />
-    <Button ghost mode="primary" on:click={() => (feedbackOpen = true)}>
-      Feedback
-    </Button>
-    {#if feedbackOpen}
-      <Portal>
-        <Modal on:close={() => (feedbackOpen = false)}>
-          <h1 slot="title">{$_("feedback.title")}</h1>
-          <UserFeedback user={$me} on:close={() => (feedbackOpen = false)} />
-        </Modal>
-      </Portal>
-    {/if}
-    {#if $$slots.action}
-      <div class="small openPane">
-        <Button
-          minW={false}
-          ghost
-          mode="primary"
-          on:click={openPanel("action")}
-        >
-          <SidebarExpand16 />
-        </Button>
-      </div>
-    {/if}
-  </header>
-
-  <main>
-    {#if $$slots.navigation}
-      <nav
-        class="navigation left"
-        class:active={panel === "navigation"}
-        id="navigation"
-      >
-        <div class="small closePane">
-          <Button ghost mode="primary" on:click={closePanel}>
-            <SidebarExpand16 />
-          </Button>
-        </div>
-        <slot name="navigation" />
-      </nav>
-    {/if}
-
-    <article class="content">
-      <slot name="content" />
-    </article>
-
-    {#if $$slots.action}
-      <nav class="action right" class:active={panel === "action"} id="action">
-        <div class="small closePane">
-          <Button ghost mode="primary" on:click={closePanel}>
-            <SidebarCollapse16 />
-          </Button>
-        </div>
-        <slot name="action" />
-      </nav>
-    {/if}
-  </main>
-
-  <div
-    class="small overlay"
-    class:active={panel !== null}
-    role="presentation"
-    on:click={closePanel}
-    on:keydown={closePanel}
-  />
-
-  <Toaster />
-</div>
-
-<style>
-  .container {
-    display: flex;
-    flex-direction: column;
-    width: 100vw;
-    height: 100vh;
-    gap: 0;
-    flex-shrink: 0;
-    background: var(--gray-1, #f5f6f7);
-  }
-
-  header {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-    background: var(--white, #ffffff);
-    border-bottom: 1px solid var(--gray-2, #d8dee2);
-    box-shadow: var(--shadow-1);
-    flex: 0 0 auto;
-    padding: 0 1rem;
-    z-index: 3;
-  }
-
-  main {
-    flex: 1 0 0;
-    display: flex;
-    flex-direction: row;
-    overflow: hidden;
-  }
-
-  nav {
-    flex: 1 0 0;
-    min-width: 16rem;
-    max-width: 18rem;
-    max-height: 100%;
-    overflow-y: auto;
-    display: flex;
-    flex-direction: column;
-    position: relative;
-    padding: 1rem;
-    gap: 1rem;
-  }
-
-  article {
-    flex: 1 0 0;
-    display: flex;
-    flex-direction: row;
-    max-height: 100%;
-    max-width: 100%;
-    overflow-y: auto;
-  }
-
-  .navigation {
-    border-right: 1px solid var(--gray-2, #d8dee2);
-  }
-
-  .action {
-    border-left: 1px solid var(--gray-2, #d8dee2);
-  }
-
-  .content {
-    padding: 0;
-  }
-
-  /* Start Mobile Styles */
-  /* When the viewport is smaller than 1024px, we move both sidebars
-  into a fixed position, then transform them to an offscreen position.
-  Then, when a button in a small-only navigation bar is clicked,
-  the appropriate sidebar will move out into a visible position. */
-
-  .small {
-    display: none;
-  }
-
-  @media (max-width: 64rem) {
-    .small {
-      display: block;
-    }
-    header {
-      gap: 0.5rem;
-      overflow-x: auto;
-    }
-    nav {
-      min-width: 33vh;
-      max-width: 100vh;
-      display: flex;
-      position: fixed;
-      top: 0;
-      bottom: 0;
-      background: var(--gray-1, #f5f6f7);
-      z-index: 2;
-      transition: transform 0.25s ease-in-out;
-      overflow: hidden;
-    }
-    nav.right {
-      transform: translateX(100%);
-      right: 0;
-      border-top-left-radius: 1rem;
-      border-bottom-left-radius: 1rem;
-    }
-    nav.left {
-      left: 0;
-      transform: translateX(-100%);
-      border-top-right-radius: 1rem;
-      border-bottom-right-radius: 1rem;
-    }
-    nav.active {
-      display: flex;
-      transform: translateX(0);
-      transition: transform 0.25s ease-in-out;
-    }
-    .small.overlay {
-      visibility: hidden;
-      position: fixed;
-      z-index: 1;
-      background: var(--gray-5, #233944);
-      top: 0;
-      left: 0;
-      right: 0;
-      bottom: 0;
-      opacity: 0;
-      transition:
-        opacity 0.25s linear,
-        visibility 0s linear 0.25s;
-    }
-    .small.overlay.active {
-      visibility: visible;
-      opacity: 0.75;
-      transition:
-        opacity 0.25s linear,
-        visibility 0s linear 0s;
-    }
-  }
-  /* End Mobile Styles */
-</style>
