+++ src/routes/+layout.svelte
@@ -6,6 +6,7 @@
 
   import { setContext } from "svelte";
   import { writable, type Writable } from "svelte/store";
+  import AppLayout from "@/lib/components/layouts/AppLayout.svelte";
 
   export let data: LayoutData;
 
@@ -30,4 +31,6 @@
   setContext("tipOfDay", data.tipOfDay);
 </script>
 
-<slot />
+<AppLayout>
+  <slot />
+</AppLayout>
