+++ src/routes/upload/+layout.svelte
@@ -1,14 +1,14 @@
 <script lang="ts">
   import "@/style/kit.css";
 
-  import MainLayout from "@/lib/components/layouts/MainLayout.svelte";
+  import SidebarLayout from "@/lib/components/layouts/SidebarLayout.svelte";
   import SignedIn from "@/lib/components/common/SignedIn.svelte";
 
   import Documents from "../documents/sidebar/Documents.svelte";
   import Projects from "../documents/sidebar/Projects.svelte";
 </script>
 
-<MainLayout>
+<SidebarLayout>
   <svelte:fragment slot="navigation">
     <Documents />
     <SignedIn>
@@ -17,4 +17,4 @@
   </svelte:fragment>
 
   <slot slot="content" />
-</MainLayout>
+</SidebarLayout>
